+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - `attributesprocessor`: Support filter by severity (#9132)
 - `processor/transform`: Add transformation of logs (#9368)
 - `datadogexporter`: Add `metrics::summaries::mode` to specify export mode for summaries (#8846)
+- `datadogexporter`: Some config validation steps are now done on `Validate` instead of `Sanitize` (#8829)
 
 ### 🧰 Bug fixes 🧰